+++ addon/views/text-area.js
@@ -1,2 +1,0 @@
-import Ember from 'ember';
-export default Ember.TextArea.extend();