+++ app/components/form-submit.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-easy-form/components/form-submit';