+++ tests/components/form-for-test.js
@@ -8,9 +8,7 @@ const {set} = Ember;
 var model;
 
 var validateFunction = function() {
-  var promise = new Ember.Deferred();
-  promise.resolve();
-  return promise;
+  return Ember.RSVP.resolve();
 };
 
 moduleForComponent('form-for', 'Integration | Component | form for', {
@@ -27,13 +25,13 @@ moduleForComponent('form-for', 'Integration | Component | form for', {
 });
 
 test('renders a form element', function(assert) {
-  this.render(hbs`{{#form-for controller}}{{/form-for}}`);
+  this.render(hbs`{{#form-for this}}{{/form-for}}`);
   assert.ok(this.$().find('form').get(0));
 });
 
 test('uses the defined wrapper', function(assert) {
   config.registerWrapper('my_wrapper', {formClass: 'my-form-class'});
-  this.render(hbs`{{#form-for controller wrapper="my_wrapper"}}{{/form-for}}`);
+  this.render(hbs`{{#form-for this wrapper="my_wrapper"}}{{/form-for}}`);
   assert.equal(this.$().find('form').attr('class'), 'ember-view my-form-class');
 });
 
@@ -71,9 +69,7 @@ test('submitting with valid model calls submit action on controller', function(a
 test('submitting with valid controller calls submit action on controller', function(assert) {
   model.reopen({
     validate: function() {
-      var promise = new Ember.Deferred();
-      promise.resolve();
-      return promise;
+      return Ember.RSVP.resolve();
     }
   });
   Ember.run(function() {
@@ -142,7 +138,7 @@ test('submitting with ember-data model without validations can call submit actio
 test('uses the specified model as the basis for {{input}} property lookup', function(assert) {
   this.set('foo', "BORING");
   this.set('theModel', { foo: "LOL" });
-  this.render(hbs`{{#form-for theModel}}{{input foo name="easy-input"}} <div id="asl">{{foo}}</div> {{input id="ember-input" value=foo}}{{/form-for}}`);
+  this.render(hbs`{{#form-for theModel}}{{input-for "foo" name="easy-input"}} <div id="asl">{{foo}}</div> {{input id="ember-input" value=foo}}{{/form-for}}`);
 
   assert.equal(this.$('input[name="easy-input"]').val(), "LOL", "easy-input uses form-for's model as its context for looking up its property");
   assert.equal(this.$('#ember-input').val(), "BORING", "vanilla ember inputs are unaffected by form-for");